Support building with msvc

+++ native/src/data_handle.cpp
@@ -5,7 +5,7 @@ bool FreeVector(std::vector<u8>* vector) {
     return true;
 }
 
-void FillData(std::vector<u8>* vector, u8** dst, u32* dst_len) {
+void FillData(std::vector<u8>* vector, u8** dst, size_t* dst_len) {
     *dst = vector->data();
     *dst_len = vector->size();
 }
